# api1-library/openapi.yaml
openapi: "3.0.3"
info:
  title: Library Management API
  version: "1.0.0"
  description: |
    도서관 도서 및 대출을 관리하는 REST API입니다.
    IBM API Connect 등록용 샘플로 설계되었으며, 인증 없이 사용 가능한
    직관적인 CRUD 패턴을 시연합니다.

    ## 주요 기능
    - 도서 등록/조회/수정/삭제 (장르·저자·가용여부 필터, 페이징 지원)
    - 도서 대출 신청 및 반납 처리
    - 재고 자동 관리 (대출 시 감소, 반납 시 증가, stock=0 시 available=false 자동 전환)

    ## 데이터 저장
    In-memory 저장소를 사용하며, 서버 재시작 시 초기 데이터로 리셋됩니다.

  contact:
    name: IBM API Connect Sample
    url: https://www.ibm.com/products/api-connect
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: 로컬 개발 서버
  - url: https://library-api.apps.your-cluster.example.com
    description: OpenShift 배포 (Route)
  - url: https://library-api.your-cluster.example.com
    description: Kubernetes 배포 (Ingress)
...
paths:
...
  /plans:
    get:
      tags: [Plans]
      summary: Plan 목록 조회
      operationId: listPlans
      description: |
        전체 Plan 목록을 조회합니다.
        특정 Product, 상태, 무료 여부로 필터링할 수 있습니다.
      parameters:
        - name: productId
          in: query
          description: 특정 Product에 속한 Plan만 조회
          schema: { type: string }
        - name: status
          in: query
          description: "배포 상태 필터 (draft | review | published | rejected | deprecated)"
          schema: { type: string, enum: [draft, review, published, rejected, deprecated] }
        - name: freeOnly
          in: query
          description: 무료 플랜(price.amount=0)만 조회
          schema: { type: boolean }
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QueryLimit"
      responses:
        "200":
          description: Plan 목록
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlanListResponse" }
    post:
      tags: [Plans]
      summary: Plan 생성
      operationId: createPlan
      description: |
        새 Plan을 **초안(draft)** 상태로 생성합니다.
        `productId`로 지정한 Product가 존재해야 합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlanInput" }
            example:
              productId: "PROD-001"
              name: "premium"
              title: "프리미엄 플랜"
              description: "무제한 호출, 전용 SLA"
              rateLimit:
                count: 10000
                interval: 1
                unit: "hour"
              price:
                amount: 99000
                currency: "KRW"
                billingPeriod: "monthly"
              features: ["전체 API", "우선 지원"]
              approvalRequired: false
      responses:
        "201":
          description: 생성된 Plan (draft 상태)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Plan" }
        "400":
          description: 필수 필드 누락 또는 유효성 오류
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: 연결할 Product를 찾을 수 없음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: 동일 name의 Plan이 같은 Product 내에 이미 존재
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /plans/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Plan ID (예 PLAN-001)
        schema: { type: string }
    get:
      tags: [Plans]
      summary: Plan 단건 조회
      operationId: getPlan
      description: 특정 Plan을 ID로 조회합니다. 연결된 Product 정보도 함께 반환합니다.
      responses:
        "200":
          description: Plan 상세 (product 포함)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlanDetail" }
        "404":
          description: Plan을 찾을 수 없음
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    put:
      tags: [Plans]
      summary: Plan 전체 수정
      operationId: replacePlan
      description: Plan을 전체 수정합니다. **draft** 또는 **rejected** 상태에서만 가능합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlanInput" }
      responses:
        "200": { description: 수정된 Plan, content: { application/json: { schema: { $ref: "#/components/schemas/Plan" } } } }
        "404": { description: Plan을 찾을 수 없음, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: "수정 불가 상태", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    patch:
      tags: [Plans]
      summary: Plan 부분 수정
      operationId: updatePlan
      description: Plan을 부분 수정합니다. **draft** 또는 **rejected** 상태에서만 가능합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlanPatchInput" }
      responses:
        "200": { description: 수정된 Plan, content: { application/json: { schema: { $ref: "#/components/schemas/Plan" } } } }
        "400": { description: 수정할 필드 없음, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Plan을 찾을 수 없음, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: "수정 불가 상태", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    delete:
      tags: [Plans]
      summary: Plan 삭제
      operationId: deletePlan
      description: Plan을 삭제합니다. **draft** 또는 **rejected** 상태에서만 가능합니다.
      responses:
        "200": { description: 삭제 완료, content: { application/json: { schema: { $ref: "#/components/schemas/PlanDeleteResponse" } } } }
        "404": { description: Plan을 찾을 수 없음, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: "삭제 불가 상태", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /plans/{id}/submit:
    post:
      tags: [Plans]
      summary: 검토 요청 (draft → review)
      operationId: submitPlanForReview
      description: Plan을 검토 요청 상태로 전환합니다. **draft** 상태에서만 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: 검토 요청 성공, content: { application/json: { schema: { $ref: "#/components/schemas/PlanActionResponse" } } } }
        "404": { description: Plan을 찾을 수 없음, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: "상태 전이 불가", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /plans/{id}/publish:
    post:
      tags: [Plans]
      summary: 게시/배포 (review → published)
      operationId: publishPlan
      description: Plan을 게시(PUBLISHED) 상태로 배포합니다. **review** 상태에서만 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: 배포 성공, content: { application/json: { schema: { $ref: "#/components/schemas/PlanActionResponse" } } } }
        "404": { description: Plan을 찾을 수 없음, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: "상태 전이 불가", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /plans/{id}/reject:
    post:
      tags: [Plans]
      summary: 거절 (review → rejected)
      operationId: rejectPlan
      description: Plan 검토를 거절합니다. **review** 상태에서만 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, description: 거절 사유 }
      responses:
        "200": { description: 거절 처리 성공, content: { application/json: { schema: { $ref: "#/components/schemas/PlanActionResponse" } } } }
        "400": { description: 거절 사유 누락, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "404": { description: Plan을 찾을 수 없음, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: "상태 전이 불가", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /plans/{id}/deprecate:
    post:
      tags: [Plans]
      summary: 사용 중단 (published → deprecated)
      operationId: deprecatePlan
      description: Plan을 사용 중단 상태로 전환합니다. **published** 상태에서만 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: 사용 중단 처리 성공, content: { application/json: { schema: { $ref: "#/components/schemas/PlanActionResponse" } } } }
        "404": { description: Plan을 찾을 수 없음, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: "상태 전이 불가", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /plans/{id}/resubmit:
    post:
      tags: [Plans]
      summary: 재제출 (rejected → draft)
      operationId: resubmitPlan
      description: 거절된 Plan을 초안 상태로 되돌립니다. **rejected** 상태에서만 가능합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: 재제출 성공, content: { application/json: { schema: { $ref: "#/components/schemas/PlanActionResponse" } } } }
        "404": { description: Plan을 찾을 수 없음, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
        "409": { description: "상태 전이 불가", content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }

# ─────────────────────────────────────────────────────────────────────────────
# Components
# ─────────────────────────────────────────────────────────────────────────────
components:
...
  schemas:
...
    RateLimitPolicy:
      type: object
      description: Rate Limit 정책
      required: [count, interval, unit]
      properties:
        count:    { type: integer, minimum: 1, description: 최대 호출 횟수 }
        interval: { type: integer, minimum: 1, description: 시간 단위 값 }
        unit:
          type: string
          enum: [second, minute, hour, day]
          description: 시간 단위
    PlanPrice:
      type: object
      description: Plan 가격 정보
      required: [amount, currency]
      properties:
        amount:
          type: number
          minimum: 0
          description: 요금 (0이면 무료)
        currency:
          type: string
          description: "통화 코드 (KRW, USD 등)"
        billingPeriod:
          type: string
          nullable: true
          description: "청구 주기 (monthly | yearly | null=무료)"
    Product:
      type: object
      description: API 상품(Product) 정보
      properties:
        id:              { type: string, description: "Product 고유 ID (예: PROD-001)" }
        name:            { type: string, description: Product 이름 (URL-safe) }
        title:           { type: string, description: Product 표시 이름 }
        description:     { type: string, description: Product 설명 }
        version:         { type: string, description: "버전 (SemVer)" }
        status:
          type: string
          enum: [draft, review, published, rejected, deprecated]
          description: 배포 상태
        visibility:
          type: string
          enum: [public, authenticated, custom]
          description: 공개 범위
        categories:      { type: array, items: { type: string }, description: 카테고리 태그 목록 }
        createdAt:       { type: string, format: date-time }
        updatedAt:       { type: string, format: date-time }
        publishedAt:     { type: string, format: date-time, nullable: true }
        reviewNote:      { type: string, nullable: true, description: 검토 메모 }
        rejectionReason: { type: string, nullable: true, description: 거절 사유 }
...
    Plan:
      type: object
      description: 사용 정책(Plan) 정보
      properties:
        id:               { type: string, description: "Plan 고유 ID (예: PLAN-001)" }
        productId:        { type: string, description: 연결된 Product ID }
        name:             { type: string, description: Plan 이름 (URL-safe) }
        title:            { type: string, description: Plan 표시 이름 }
        description:      { type: string, description: Plan 설명 }
        rateLimit:        { $ref: "#/components/schemas/RateLimitPolicy" }
        burstLimit:
          $ref: "#/components/schemas/RateLimitPolicy"
        price:            { $ref: "#/components/schemas/PlanPrice" }
        features:         { type: array, items: { type: string }, description: 허용 기능 목록 }
        status:
          type: string
          enum: [draft, review, published, rejected, deprecated]
        approvalRequired: { type: boolean, description: 구독 시 관리자 승인 필요 여부 }
        createdAt:        { type: string, format: date-time }
        updatedAt:        { type: string, format: date-time }
    PlanDetail:
      allOf:
        - $ref: "#/components/schemas/Plan"
        - type: object
          properties:
            product: { $ref: "#/components/schemas/Product" }
    PlanInput:
      type: object
      description: Plan 생성/수정 입력
      required: [productId, name, title, description, rateLimit, price]
      properties:
        productId:        { type: string, description: 연결할 Product ID }
        name:             { type: string, description: Plan 이름 (URL-safe) }
        title:            { type: string }
        description:      { type: string }
        rateLimit:        { $ref: "#/components/schemas/RateLimitPolicy" }
        burstLimit:
          $ref: "#/components/schemas/RateLimitPolicy"
        price:            { $ref: "#/components/schemas/PlanPrice" }
        features:         { type: array, items: { type: string } }
        approvalRequired: { type: boolean, default: false }
    PlanPatchInput:
      type: object
      description: Plan 부분 수정 입력
      properties:
        title:            { type: string }
        description:      { type: string }
        rateLimit:        { $ref: "#/components/schemas/RateLimitPolicy" }
        burstLimit:
          $ref: "#/components/schemas/RateLimitPolicy"
        price:            { $ref: "#/components/schemas/PlanPrice" }
        features:         { type: array, items: { type: string } }
        approvalRequired: { type: boolean }
    PlanListResponse:
      type: object
      properties:
        total:      { type: integer }
        page:       { type: integer }
        limit:      { type: integer }
        totalPages: { type: integer }
        data:       { type: array, items: { $ref: "#/components/schemas/Plan" } }
    PlanActionResponse:
      type: object
      properties:
        message: { type: string }
        plan:    { $ref: "#/components/schemas/Plan" }
    PlanDeleteResponse:
      type: object
      properties:
        message:     { type: string }
        deletedPlan: { $ref: "#/components/schemas/Plan" }
    ErrorResponse:
      type: object
      properties:
        error:  { type: string, description: 에러 메시지 }
        status: { type: string, nullable: true }
        id:     { type: string, nullable: true }

  # ── Parameters ───────────────────────────────────────────────────────────────
  parameters:
...
    QueryPage:
      name: page
      in: query
      required: false
      description: 페이지 번호 (1부터 시작)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    QueryLimit:
      name: limit
      in: query
      required: false
      description: 페이지당 결과 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      example: 10
...